+++ build.ps1
@@ -5,9 +5,9 @@ If (Test-Path "${PSScriptRoot}\artifacts") {Remove-Item -Recurse -Force "${PSScr
 
 # Build for Windows, macOS and Linux
 # Use linker flags for shrinking
-set GOOS=linux
+Set-Variable GOOS=linux
 go build -o "${PSScriptRoot}\artifacts\serve_Windows_x64.exe" -ldflags="-s -w" "github.com/philippgille/serve"
-set GOOS=linux
+Set-Variable GOOS=linux
 go build -o "${PSScriptRoot}\artifacts\serve_macOS_x64" -ldflags="-s -w" "github.com/philippgille/serve"
-set GOOS=linux
+Set-Variable GOOS=linux
 go build -o ${PSScriptRoot}\"artifacts\serve_Linux_x64" -ldflags="-s -w" "github.com/philippgille/serve"